+++ src/pipeline/highp.rs
@@ -116,9 +116,16 @@ pub const STAGES: &[StageFn; super::STAGES_COUNT] = &[
     xy_to_radius,
     xy_to_2pt_conical_focal_on_circle,
     xy_to_2pt_conical_well_behaved,
+    xy_to_2pt_conical_smaller,
     xy_to_2pt_conical_greater,
+    xy_to_2pt_conical_strip,
+    mask_2pt_conical_nan,
     mask_2pt_conical_degenerates,
     apply_vector_mask,
+    alter_2pt_conical_compensate_focal,
+    alter_2pt_conical_unswap,
+    negate_x,
+    apply_concentric_scale_bias,
     gamma_expand_2,
     gamma_expand_dst_2,
     gamma_compress_2,
@@ -1017,16 +1024,34 @@ fn xy_to_2pt_conical_greater(p: &mut Pipeline) {
     p.next_stage();
 }
 
-fn mask_2pt_conical_degenerates(p: &mut Pipeline) {
-    let ctx = &mut p.ctx.two_point_conical_gradient;
+// From: https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/src/opts/SkRasterPipeline_opts.h;l=3625;drc=d5c02fd369596fb4a6fa563f5506ff71ff7299ed
+fn xy_to_2pt_conical_smaller(p: &mut Pipeline) {
+    let ctx = &p.ctx.two_point_conical_gradient;
 
-    let t = p.r;
-    let is_degenerate = t.cmp_le(f32x8::default()) | t.cmp_ne(t);
-    p.r = is_degenerate.blend(f32x8::default(), t);
+    let x = p.r;
+    let y = p.g;
+    p.r = -(x * x - y * y).sqrt() - x * f32x8::splat(ctx.p0);
 
+    p.next_stage();
+}
+
+// From: https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/src/opts/SkRasterPipeline_opts.h;l=3605;drc=d5c02fd369596fb4a6fa563f5506ff71ff7299ed
+fn xy_to_2pt_conical_strip(p: &mut Pipeline) {
+    let ctx = &p.ctx.two_point_conical_gradient;
+
+    let x = p.r;
+    let y = p.g;
+    p.r = x + (f32x8::splat(ctx.p0) - y * y).sqrt();
+
+    p.next_stage();
+}
+
+
+#[inline]
+fn mask_2pt_conical_from_degenerate(is_degenerate: f32x8) -> u32x8 {
     let is_not_degenerate = !is_degenerate.to_u32x8_bitcast();
     let is_not_degenerate: [u32; 8] = bytemuck::cast(is_not_degenerate);
-    ctx.mask = bytemuck::cast([
+    bytemuck::cast([
         if is_not_degenerate[0] != 0 { !0 } else { 0 },
         if is_not_degenerate[1] != 0 { !0 } else { 0 },
         if is_not_degenerate[2] != 0 { !0 } else { 0 },
@@ -1035,7 +1060,29 @@ fn mask_2pt_conical_degenerates(p: &mut Pipeline) {
         if is_not_degenerate[5] != 0 { !0 } else { 0 },
         if is_not_degenerate[6] != 0 { !0 } else { 0 },
         if is_not_degenerate[7] != 0 { !0 } else { 0 },
-    ]);
+    ])
+}
+
+
+// From: https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/src/opts/SkRasterPipeline_opts.h;l=3641;drc=d5c02fd369596fb4a6fa563f5506ff71ff7299ed
+fn mask_2pt_conical_nan(p: &mut Pipeline) {
+    let ctx = &mut p.ctx.two_point_conical_gradient;
+
+    let t = p.r;
+    let is_degenerate = t.cmp_ne(t);
+    p.r = is_degenerate.blend(f32x8::default(), t);
+    ctx.mask = mask_2pt_conical_from_degenerate(is_degenerate);
+
+    p.next_stage();
+}
+
+fn mask_2pt_conical_degenerates(p: &mut Pipeline) {
+    let ctx = &mut p.ctx.two_point_conical_gradient;
+
+    let t = p.r;
+    let is_degenerate = t.cmp_le(f32x8::default()) | t.cmp_ne(t);
+    p.r = is_degenerate.blend(f32x8::default(), t);
+    ctx.mask = mask_2pt_conical_from_degenerate(is_degenerate);
 
     p.next_stage();
 }
@@ -1047,6 +1094,40 @@ fn apply_vector_mask(p: &mut Pipeline) {
     p.g = (p.g.to_u32x8_bitcast() & ctx.mask).to_f32x8_bitcast();
     p.b = (p.b.to_u32x8_bitcast() & ctx.mask).to_f32x8_bitcast();
     p.a = (p.a.to_u32x8_bitcast() & ctx.mask).to_f32x8_bitcast();
+
+    p.next_stage();
+}
+
+// From: https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/src/opts/SkRasterPipeline_opts.h;l=3630;drc=d5c02fd369596fb4a6fa563f5506ff71ff7299ed
+fn alter_2pt_conical_compensate_focal(p: &mut Pipeline) {
+    let ctx = &p.ctx.two_point_conical_gradient;
+
+    p.r = p.r + f32x8::splat(ctx.p1);
+
+    p.next_stage();
+}
+
+// From: https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/src/opts/SkRasterPipeline_opts.h;l=3636;drc=d5c02fd369596fb4a6fa563f5506ff71ff7299ed
+fn alter_2pt_conical_unswap(p: &mut Pipeline) {
+    p.r = f32x8::splat(1.0) - p.r;
+
+    p.next_stage();
+}
+
+// From: https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/src/opts/SkRasterPipeline_opts.h;l=3603;drc=d5c02fd369596fb4a6fa563f5506ff71ff7299ed
+fn negate_x(p: &mut Pipeline) {
+    p.r = -p.r;
+
+    p.next_stage();
+}
+
+// Applies the matrix operation from: https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/src/shaders/gradients/SkConicalGradient.cpp;l=206;drc=075316994c97ee86961b369bb2bff246aaa9d6c4
+fn apply_concentric_scale_bias(p: &mut Pipeline) {
+    let ctx = &p.ctx.two_point_conical_gradient;
+
+    // Apply t = t * scale + bias for concentric gradients
+    let x = p.r;
+    p.r = x * f32x8::splat(ctx.p0) + f32x8::splat(ctx.p1);
 
     p.next_stage();
 }