+++ src/shaders/mod.rs
@@ -4,6 +4,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+mod conical_gradient;
 mod gradient;
 mod linear_gradient;
 mod pattern;
@@ -11,6 +12,7 @@ mod radial_gradient;
 
 use tiny_skia_path::{NormalizedF32, Scalar};
 
+pub use conical_gradient::ConicalGradient;
 pub use gradient::GradientStop;
 pub use linear_gradient::LinearGradient;
 pub use pattern::{FilterQuality, Pattern, PixmapPaint};
@@ -52,6 +54,8 @@ pub enum Shader<'a> {
     LinearGradient(LinearGradient),
     /// A radial gradient shader.
     RadialGradient(RadialGradient),
+    /// A conical gradient shader.
+    ConicalGradient(ConicalGradient),
     /// A pattern shader.
     Pattern(Pattern<'a>),
 }
@@ -63,6 +67,7 @@ impl Shader<'_> {
             Shader::SolidColor(ref c) => c.is_opaque(),
             Shader::LinearGradient(ref g) => g.is_opaque(),
             Shader::RadialGradient(_) => false,
+            Shader::ConicalGradient(_) => false,
             Shader::Pattern(_) => false,
         }
     }
@@ -80,6 +85,7 @@ impl Shader<'_> {
             }
             Shader::LinearGradient(ref g) => g.push_stages(cs, p),
             Shader::RadialGradient(ref g) => g.push_stages(cs, p),
+            Shader::ConicalGradient(ref g) => g.push_stages(cs, p),
             Shader::Pattern(ref patt) => patt.push_stages(cs, p),
         }
     }
@@ -92,6 +98,9 @@ impl Shader<'_> {
                 g.base.transform = g.base.transform.post_concat(ts);
             }
             Shader::RadialGradient(g) => {
+                g.base.transform = g.base.transform.post_concat(ts);
+            }
+            Shader::ConicalGradient(g) => {
                 g.base.transform = g.base.transform.post_concat(ts);
             }
             Shader::Pattern(p) => {
@@ -122,6 +131,9 @@ impl Shader<'_> {
                 g.base.apply_opacity(opacity);
             }
             Shader::RadialGradient(g) => {
+                g.base.apply_opacity(opacity);
+            }
+            Shader::ConicalGradient(g) => {
                 g.base.apply_opacity(opacity);
             }
             Shader::Pattern(ref mut p) => {