+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- 2-point conical gradient support (`ConicalGradient`).
+
 
 ## [0.11.4] - 2024-02-04
 ### Fixed